pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "humanoid-robotics-book"
version = "0.1.0"
description = "A project for teaching Physical AI & Humanoid Robotics with RAG capabilities"
readme = "README.md"
authors = [
    {name = "Project Author"}
]
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]

dependencies = [
    "fastapi>=0.104.1,<1.0.0",
    "uvicorn[standard]>=0.24.0,<1.0.0",
    "pydantic>=2.5.0,<3.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "python-multipart>=0.0.6,<1.0.0",
    "qdrant-client>=1.10.0,<2.0.0",
    "sqlalchemy>=2.0.23,<3.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "pyjwt>=2.8.0,<3.0.0",
    "requests>=2.31.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "cohere>=5.0.0,<6.0.0",
    "google-generativeai>=0.3.1,<1.0.0",
    "sentence-transformers>=5.1.2,<6.0.0",
    "alembic>=1.13.1,<2.0.0",
    "pytest>=7.4.3,<8.0.0",
    "httpx>=0.25.2,<1.0.0",
    "pytest-asyncio>=0.21.1,<1.0.0",
    "pylint>=3.0.2,<4.0.0",
    "black>=23.11.0,<24.0.0",
    "mypy>=1.7.1,<2.0.0"
]

[project.optional-dependencies]
dev = [
    "black>=23.11.0,<24.0.0",
    "mypy>=1.7.1,<2.0.0",
    "pylint>=3.0.2,<4.0.0",
    "pytest>=7.4.3,<8.0.0",
    "pytest-asyncio>=0.21.1,<1.0.0"
]

[project.urls]
Homepage = "https://github.com/your-username/humanoid-robotics-book"
Repository = "https://github.com/your-username/humanoid-robotics-book"

[tool.setuptools]
packages = ["backend", "backend.api", "backend.models", "backend.utils", "backend.config", "backend.services"]

[tool.setuptools.package-dir]
"backend" = "backend"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests", "backend/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra"
]